+++ modules/components/ext/controllers/ComboController.mjs
@@ -2,6 +2,7 @@
  * Copyright (C) 2015, 2025; Green Screens Ltd.
  */
 
+import { GSAttr } from "../../../base/GSAttr.mjs";
 import { InteractiveController } from "./InteractiveController.mjs";
 
 
@@ -19,4 +20,26 @@ export class ComboController extends InteractiveController {
     return this.component;
   }
 
+  onChange(e) {  
+    super.onChange?.(e);
+    this.#onAttributeHandler(e);
+  }
+
+  onBlur(e) {
+    super.onBlur?.(e);
+    this.#onAttributeHandler(e);
+  }
+
+  #onAttributeHandler(e) {
+    const me = this;
+    if (!me.component.toggling) return;
+    const target = me.component.dataset.gsfTarget;
+    const value = me.component.dataset.gsfValue;
+    const attr = me.component.dataset.gsfAttribute;
+    const flag = me.component.value === value;
+    me.formElements
+      .filter(el => el != me.component)
+      .filter(el => target ? el.matches(target) : true)
+      .forEach(el => GSAttr.toggle(el, attr, flag));
+  }
 }